将 EgeControlBase 移出 ege.h
maintest.cpp 中后半部分注释就是 egenet.cpp，故删去
