+++ src/ege/fps.h
@@ -5,6 +5,8 @@
 #error include "fps.h" must after include "ege.h" or "graphics.h"
 #endif
 
+#include "egecontrolbase.h"
+
 namespace ege {
 
 class fps : public egeControlBase
